requirements.txt
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.18.2
Pillow==7.1.1
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
